% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCA_wrappers.r
\name{qcPlot}
\alias{qcPlot}
\title{Make QC plots}
\usage{
qcPlot(anno, dendcluster_anno, name, scaleLimits = c(-5000, 12000),
  scaleBreaks = seq(0, 12000, 2000), scaleLabels = seq(0, 12, 2),
  ylab = "value", fileName = gsub("\\\\.", "_", gsub("_label", "",
  name)), outputFolder)
}
\arguments{
\item{anno}{Annotation data frame for patch-seq}

\item{dendcluster_anno}{what to cluster by}

\item{scaleLimits}{scaleLimits}

\item{scaleBreaks}{scaleBreaks}

\item{scaleLabels}{scaleLabels}

\item{ylab}{ylab}

\item{fileName}{fileName}

\item{groupvars}{what to group by}
}
\value{
the plot is returned
}
\description{
Makes QC plots
}
