% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCA_wrappers.r
\name{reorder_matrix}
\alias{reorder_matrix}
\title{Reorder a matrix}
\usage{
reorder_matrix(matrix1, by.rows = TRUE)
}
\arguments{
\item{matrix1}{a matrix (rows=genes x columns=samples) of gene expression data (e.g., scRNA-seq)}

\item{by.rows}{By rows (TRUE; default) or by columns}
}
\value{
a reordered matrix
}
\description{
This function reorders a matrix by rows of columns
}
