% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCA_wrappers.r
\name{summarySE}
\alias{summarySE}
\title{Get some summary statistics}
\usage{
summarySE(data = NULL, measurevar, groupvars = NULL, na.rm = FALSE,
  conf.interval = 0.95, .drop = TRUE, roundall = F)
}
\arguments{
\item{data}{Annotation data frame}

\item{measurevar}{what variables to measure}

\item{groupvars}{what to group by}

\item{na.rm}{how to treat NA}

\item{conf.interval}{confidence interval (default = .95)}

\item{.drop}{drop something?}

\item{roundall}{should everything be rounded}
}
\value{
a data frame of statistics
}
\description{
This does the summary. For each group return a vector with N, mean, and sd.
This is called by qcPlot.
}
